RecordDetailActivitys' navigateUp() method fixed, minimum birthdate year changed from hardcode
